+++ contracts/token/ERC20/libraries/ERC20Storage.sol
@@ -19,8 +19,8 @@ library ERC20Storage {
         uint256 totalSupply;
     }
 
-    bytes32 public constant ERC20_STORAGE_POSITION = bytes32(uint256(keccak256("animoca.token.ERC20.ERC20.storage")) - 1);
-    bytes32 public constant ERC20_VERSION_SLOT = bytes32(uint256(keccak256("animoca.token.ERC20.ERC20.version")) - 1);
+    bytes32 public constant ERC20_STORAGE_POSITION = bytes32(uint256(keccak256("animoca.core.token.ERC20.ERC20.storage")) - 1);
+    bytes32 public constant ERC20_VERSION_SLOT = bytes32(uint256(keccak256("animoca.core.token.ERC20.ERC20.version")) - 1);
 
     event Transfer(address indexed from, address indexed to, uint256 value);
     event Approval(address indexed owner, address indexed spender, uint256 value);